refactor: reshape cli modules
