+++ src/rust/cli/src/execution/mod.rs
@@ -1,0 +1,9 @@
+mod authorization;
+mod orchestrator;
+mod settings;
+mod validation;
+
+pub use orchestrator::{execute, execute_with_observer, should_attempt_azure_auth};
+
+#[cfg(test)]
+mod tests;
