+++ src/rust/cli/src/main.rs
@@ -10,14 +10,8 @@ use std::{ffi::OsString, time::Instant};
 use ui::CliPresenter;
 
 fn main() {
-    let mut raw_args: Vec<OsString> = std::env::args_os().collect();
-    if raw_args.len() == 1 {
-        raw_args.push(OsString::from("--help"));
-    }
-
-    let matches = build_command().get_matches_from(raw_args.clone());
-    let args = CliArgs::from_arg_matches(&matches)
-        .expect("clap should only return matches that satisfy CliArgs");
+    let raw_args = raw_args_with_help();
+    let args = parse_args(&raw_args);
     let mut telemetry = TelemetryRecorder::from_cli_args(&raw_args, &args, NoopTelemetrySink);
     let started_at = Instant::now();
     let mut presenter = CliPresenter::detect();
@@ -34,4 +28,19 @@ fn main() {
             std::process::exit(1);
         }
     }
+}
+
+fn raw_args_with_help() -> Vec<OsString> {
+    let mut raw_args: Vec<OsString> = std::env::args_os().collect();
+    if raw_args.len() == 1 {
+        raw_args.push(OsString::from("--help"));
+    }
+
+    raw_args
+}
+
+fn parse_args(raw_args: &[OsString]) -> CliArgs {
+    let matches = build_command().get_matches_from(raw_args.iter().cloned());
+    CliArgs::from_arg_matches(&matches)
+        .expect("clap should only return matches that satisfy CliArgs")
 }
