+++ src/rust/cli/src/telemetry/mod.rs
@@ -1,0 +1,11 @@
+mod events;
+mod recorder;
+mod redaction;
+mod sink;
+
+pub use events::{ErrorEvent, FeatureUsageEvent, TelemetryContext, TelemetryEvent};
+pub use recorder::TelemetryRecorder;
+pub use sink::{MemoryTelemetrySink, NoopTelemetrySink, TelemetrySink};
+
+#[cfg(test)]
+mod tests;
