+++ src/rust/cli/src/ui/mod.rs
@@ -1,0 +1,5 @@
+mod format;
+mod presenter;
+mod render;
+
+pub(crate) use presenter::CliPresenter;
